requirements.txt
playwright
streamlit
google-generativeai
beautifulsoup4
python-dotenv
